[YDB#758] New r134/ydb758 test for %PEEKBYNAME() option node_local.max_procs
This commit adds a new r134/ydb758 test that tests that a %PEEKBYNAME() on `node_local.max_procs` works and returns an accurate process count. The test first does a `%PEEKBYNAME("node_local.max_procs",<region)` on a newly created database to verify that the most processes accessing the database is 1. Then, it creates 10 background processes that run concurrently, waits for them to finish and checks that the new maximum process count is 11 (including the parent process in the foreground that starts the background processes). Finally, it does a `DSE DUMP -FILEHEADER` to verify that this shows up in the output and a `DSE CHANGE -FILEHEADER -RESET_MAX_PROCS` to verify that resetting the maximum works as expected.

This commit also updates reference files for existing tests that do a `DSE DUMP -FILEHEADER` to expect the 2 new fields in the output.