[#403] New v63011/gtm8398 test for segment extension_count > 65355 blocks
In v6.3-011, the limit for a segment's `extension_count` field was increased from 65355 blocks to 1,048,575 blocks. This test verifies that this limit has been increased by setting the DEFAULT segment's `extension_count` to a value larger than 65355 (70000) and then doing a `MUPIP EXTEND DEFAULT` to extend the database by the new maximum of 1,048,575 blocks. This produces a `GDE-I-VALTOOBIG` error on pre-V6.3-011 builds and then the `MUPIP EXTEND DEFAULT` extends the database by the default `extension_count` value (100 blocks). The test also attempts to set `extension_count` to 1,048,576 blocks (1 more than the maximum) to verify that doing this produces the expected `GDE-I-VALTOOBIG` error.