Fix rare msreplic_F/instance_file_simple failure by waiting for connection to be established before doing updates
This commit fixes a rare failure in the msreplic_F/instance_file_simple test. This failure occurs because the test establishes connections to 2 receiver servers in the background and then continues. A few lines later, it performs 70 transactions. If this line executes before the connections are established, the `Connect Sequence Number` will be 71 instead of 1, resulting in a test failure. This commit adds 2 waits so that the connections from INST1 to INST2 and INST3 are guaranteed to be established before INST1 runs its updates.