[#406] New v63011/gtm9215 test ensures that writing ISVs that can be evaluated as exponential numbers works
This commit adds a new v63011/gtm9215 test that ensures that intrinsic special variables (ISVs) set to a string value that starts with `1E` and is followed by numbers are evaluated as strings not exponential numbers. Prior to upstream version V6.3-011, they would be evaluated as exponential numbers which would result in a `%YDB-E-NUMOFLOW` error message on a subsequent write for values of `1E47` or larger but the values would still show up in a `zshow "I"`.