[#409] New v63011/gtm9226 test ensures $translate sets output to correct length
This commit adds a v63011/gtm9226 test that checks that the output of $translate when it is passed a malformed character has the correct length. Prior to V6.3-011, there was a bug where the length could be incorrect and possibly result in SIG-11s or other issues with subsequent code.