Exclude r130/ydb534 subtest on Ubuntu 21.10 due to what we suspect is a Ubuntu regression
In Ubuntu 21.10, a call to abort(), instead of just sending a signal-6 to the process also started
creating an unrequested core dump even though the signal was handled. This then created two core
files (the handling of the sig-6 intentionally created a core using sig-3) which the test did not
expect. This caused a test failure. The simplest solution is to bypass this subtest on Ubuntu 21.10
and hoping 22.04 holds a fix. This change is only looking explicitly for Ubuntu 21.10 so this
subtest will start running again when subsequent versions become available.