[#399] Add defer xxx.Free() calls to example Go routines
Although this issue was primarily written against wordfreq.go, that routine was already fixed up
in an October 2021 commit:

commit 4d5a32c138560630d3bcb505b2179b6ce81280f9
Author: Steven Estes <steven.estes@yottadb.com>
Date:   Tue Oct 26 11:21:11 2021 -0400

    Add missing defers for release of buffers and strings used in wordfreq test

Strictly speaking, the defer xxx.Free() is not required for correct programming as the C memory
anchored within these Go structures like BufferT is automatically released when the Go structure
goes out of scope and is garbage collected. Nevertheless, YottaDB recommends explicit DEFER
commands to release the storage when done with it for better storage management - especially if
the allocated buffers are large.