Fix test failures due to tcsh 6.23.02 sometimes interpreting '\' followed by a number as a special character
Several tests failed on our in-house arch machine due to a recent tcsh update that now interprets some character sequences involving `\` followed by a number as special characters. As both the old and new versions of tcsh interpret `\\` as a special character sequence for `\`, the fix is to replace `\` characters that are incorrectly interpreted as escape characters with `\\`.

The r120/divzero test failed due to the `\1` sequence of characters being interpreted as an invisible special character which caused the `\1` to disappear.

The v53002/C9I06003000 test failed because this tcsh change caused `$record` to be set incorrectly. The test sets `$record` to record 0x200000 from the journal extract. Part of this record was interpreted as `\3,68264948\586028059>93>93\^a=` instead of `\66143,68264\7948\711458602\38059\7693\7693\^a=`. In this case, the fix was to pass the output of the `grep` command to `sed` which changes all instances of `\` to `\\` ensuring that `$updatenumber` is set correctly later on in the test.

The output of `tcsh --version` is `tcsh 6.23.02 (Astron) 2022-01-03 (x86_64-unknown-linux) options wide,nls,dl,al,kan,sm,rh,color,filec`.