[YottaDB/Lang/YDBGo#37] Add test that exit from Go wrapper does not have a 5 second delay
The exact conditions under which this happened are not crystal clear. This issue was reported after
it had already been fixed (as part of [YottaDB/Lang/YDBGo#34]). The case that was happening previously
to the author was a signal handler had to be active at the time - meaning the failure was dependent
on a development version of the YottaDB/Lang/YDBGo#34 change that has since been withdrawn so between
that and the fix that was done, there is no known way to create this error at this time (even on
prior released versions of the wrapper and/or YottaDB).

But this test reconstructs how they were occurring previously by getting a notification of a signal
(meaning the signal is in-process when we run yottadb.Exit()) so if it shows up again, this test should
catch it.