Fix sudo subtests to return the actual status code when the status code is non-zero
The sudo/plugins test failed with a non-zero status in internal testing. The error message from the failure read `ydbinstall returned a non-zero status: 0` incorrectly stating that ydbinstall had returned a non-zero status of `0`.

The cause of this failure was a bug in the test script. To check if ydbinstall had returned a non-zero status code, the test did the following:

```
if [ 0 != $? ]; then
        echo "ydbinstall returned a non-zero status: $?"
```

The problem with the above lies in the second line. The `if [ 0 != $? ]` check will work correctly because the value of `$?` will be the status code from `ydbinstall`. However, the value of `$?` on the following line is the return code of the `if` command not the return code of the `ydbinstall` command. This means that it will always incorrectly claim that `ydbinstall returned a non-zero status: 0`.

The fix for this is to set the return code of ydbinstall to a variable `$status` and check the value of `$status` not `$?`. Since this bug is present in the other sudo subtests, this fix has been applied to them as well.