[YottaDB/Lang/YDBGo#34] Additional cleanup changes
Also add ydbgo34d executable to test syslogEntry() routine and sending messages to syslog.

Change the "message to fetch" in randomWalk.c to a real error. Previous it would fetch message id -1 which would cause
it to return a not found error but now, if the message has an invalid facility id, it will panic and that started making
this -1 error fetch panic so gave it a real error instead (yottadb.YDB_ERR_KILLBYSIGUINFO).