[YDB#603] [V63012] Update test reference files to reflect V6.3-012 changes and remove obsolete v44004/C9E03002545 test
The following tests have been updated to reflect changes to the YottaDB codebase after merging in V6.3-012:

Required updates to the reference file due to the addition of the new %JSWRITE routine (#421):
- `mpt_extra/rtn`
- `r134/ydb781`

Required updates to the reference file due to the additions of the `Full Block Write` field in DSE output (#415):
- `dse/dse_chng_fhead`
- `jnl/jnl_set`
- `mupip/mu_integ`
- `nullsubs_coll/call_zwr_go`
- `nullsubs_coll/gde_com_nullsub`
- `nullsubs_coll/multipledb_binload`
- `nullsubs_coll/multipledb_binxtract`
- `nullsubs_coll/stdnullcoll`

Required updates to the reference file due to changes (#415) that added a `FBWR` field to the GDE segment output:
- `gde/basic`: Required updates to the reference files `dmi.txt`, `gdemaptst.txt`, `its.txt` and `tnt.txt` due to these changes
- `gde/gblname`
- `gde/longname`: Also required updates to the reference file `dbcreate1.txt`
- `gde/misc`
- `gde/name_subscripts`
- `gde/showregion`: Required updates because the addition of the `FBWR` field changed line numbers for GDE SHOW output
- `jnl/jnl_logical`
- `jnl/recover`
- `nullsubs_coll/gde_com_nullsub`
- `nullsubs_coll/stdnullcoll
- `spanning_nodes/gtm6941`
- `unicode/filenames`
- `v51000/D9G01002589`
- `v51000/D9G01002590`

Required updates to the reference file due to the addition of FULLBLKWRT to GDE SHOW (#415):
- `r126/ydb429`

Required updates to the reference file due to change to error code for `YDB_ERR_ZGBLDIRACC` error:
- `errors/err_messages`
- `v43001d/D9C08002194`
- `v53002/C9C11002181`
- `v63003/gtm8844`
- `r124/ydb397`

Required updates to the reference file due to the additions of the new `MUPIP INFO` messages:
- `64bittn/mupip_integ`
- `dse/dse_chng_fhead`
- `v53003/C9I06002997`
- `v53003/D9H12002672`

Removed dbcertify parts of test since dbcertify was removed in V6.3-012:
`v53003/D9I10002703`: Removed dbcertify from test and from `errors_*.txt` reference files.

Misc:
- `v44004/C9E03002545`: Removed test as it tests the $gtm_fullblockwrite/$ydb_fullblockwrite variable that was removed by #415.
  It will be replaced by a new v63012/gtm5381 test.
- `v53003/D9I10002703`: Removed $gtm_full_block_writes/$ydb_full_block_writes and added
  errors_gtm_string_pool_limit.txt/errors_ydb_string_pool_limit.txt reference files due to #420 changes to string pool limits
- `v63005/gtm8877`: Updated reference file to expect 2 more YDB_E_RESTRICTEDOP messages

This commit also enables the v63012 test suite in `COM/SUITE`.