Combine sudo/plugins and sudo/pluginsonly tests and disable sudo/plugins test on ARMVXL due to periodic failures involving network issues on ARMV6L
The sudo/plugins test has been failing on ARM with various network connections errors at a rate of about once a week. The test and the related sudo/pluginsonly test verify that 30 different combinations of plugins install successfully via YottaDB.

The failure messages have resembled either of the following:

```
fatal: unable to access 'https://gitlab.com/YottaDB/Util/YDBAIM.git/': gnutls_handshake() failed: The TLS connection was non-properly terminated.
curl: (52) Empty reply from server
```

So far, these failures have involved 1 plugin install out of all the plugin installs in the test. An attempted fix (YDBTest!1350) was to do a retry of a failed plugin install on ARMVXL. In internal testing, this did not fix the failures so this approach was abandoned.

This commit disables the sudo/plugins test on ARMVXL. Running the test on x86_64 and AARCH64 is sufficient to verify that ydbinstall is installing plugins successfully.

This commit also changes the ydbinstall commands to output stderr into the same file as the ydbinstall output. This change was retained from YDBTest!1350 as the messages about setting UTF-8 locales are irrelevant to the purpose of this test and confusing when the test fails and they appear out of order in the test output. The reference file is updated to reflect this change.

Since the sudo/pluginsonly test shares a large amount of code with the sudo/plugins test, the 2 tests have been combined into a single sudo/plugins test that tests installing plugins with and without the `--plugins-only` ydbinstall option.