Fix cargo command not found; add missing clang
Due to
https://gitlab.com/YottaDB/DB/YDB/-/commit/1607247a0ad118bec7c242712984f42492c351f3,
we no longer use the system provided Rust, but rather install the stable
version. However, the stable version installed by the "rustup" tool is
installed in $HOME/.cargo/bin; not in any of the standard places.

Per rustup's documentation, you always need to add it in $PATH.

https://github.com/alexcrichton/rustup.rs

Clang is required for the Rust Plugin. It was removed in
https://gitlab.com/YottaDB/DB/YDB/-/commit/1607247a0ad118bec7c242712984f42492c351f3,
but I will add it back as we need it, but keeping it here for
completeness' sake is okay.