[#421] New v63012/gtm9244 test for %JSWRITE utility routine
This commit adds a v63012/gtm9244 test that verifies that the %JSWRITE utility routine works correctly. It does this by creating a database, filling it with globals
and testing various combinations of inputs to `%JSWRITE` to ensure that it returns the appropriate JSON output. The output is written to a file and then checked by
an M routine to verify that the correct variables were written by `%JSWRITE` and that the values of the variables in the database are the same as their values in
the `JSWRITE` output. Finally, the test verifies that it returns appropriate error messages given incorrect input.