[#427] New v63012/gtm9269 test that -nowarning suppresses %YDB-W-DONOBLOCK messages
This commit adds a v63012/gtm9269 test that verifies that the `-nowarning` compilation qualifier suppresses %YDB-W-DONOBLOCK messages.
A %YDB-W-DONOBLOCK message appears when the compiler detects an argumentless `do` not followed by a block and optimizes it out of the
code. This is almost always a logic error. Prior to V6.3-012, `-nowarning` did not suppress this warning message.