[#430] Add return code check from setupgoenv.csh to several tests that weren't checking it
We have periodic failures attempting to do a go get for the yottadb package. Although setupgoenv.csh has a
retry, sometimes it fails 5 times and gives up. But when tests don't check the return code and continue to
attempt to run, it makes for a big mess of a failure log that can take some time to diagnose so checking if
the wrapper was successfully downloaded and installed just makes everything better. The 5 scripts updated in
this MR were missing that check.

Also fixed the failing script's location in the error 'echo' statement which had the wrong path for the script.