Fix r130/ydb534 reference file due to changes in core trace in newer Linuxen.
This was an issue encountered on Arch. The test creates an expected core file and
looks to make sure the intentional abort in the test code was what caused it. The
extracted message from the core traceback has a stack-level number associated with it.
Due to recent changes on Arch, the stack level increased by 1. Example follows:

On the Arch box the traceback looked like:

On the Ubuntu 20.04 box the traceback looked like:

Note that frame 5 in the Arch box back trace is not reflected in the Ubuntu trace back. The fix is to allow the do_abort line to be for frame 7 or 8.