Ensure that sudo test does not overwrite YottaDB's pkg-config file
This commit modifies the instream of the sudo test to backup YottaDB's pkg-config file and restore it after the test
completes. Without this change, the pkg-config file will be overwritten to point to the sudo test's version if the
test is run with the latest version on the system. The backup is done before the start of the first subtest and the
restore is done after the last subtest ends to ensure that nothing is permanently altered.