Fix rare r134/ydb546 subtest failure when run with a Debug build due to deeply nested boolean expressions
The r134/ydb546 test failed in in-house testing. The cause of the failure was a
boolean expression that was nested to a depth of 4 which is above the maximum
allowed by default. This boolean expression was in the ydb546booltest routine
which generates a boolean expression that gets longer as the test continues
until a `maxwait` has elapsed. Raising the nesting limit to 8 ensures that
the test will pass successfully without failing an assert.

The assert failure was:

```
%YDB-F-ASSERT, Assert failed in sr_port/bool_init.c line 64 for expression (max_boolexpr_nesting_depth > boolZysqlnull->boolexpr_nesting_depth)
```

The generated boolean expression that caused it was:

```
set xstr="set:(($test='$test)'!((true&((((('false)'[('false))&$test)'!((1&(('true)']]true))<=false))'?1""1""))'?1""1"")) result=1"
```