Add v63012/missedrevert test
After fixing the clang compiler warning in the `dqnoop()` macro as part of YDB!1132,
we noticed an issue with strange SIG-11s in `FOR` commands. This was due to a
silent merge conflict in `sr_port/m_for.c` (see https://gitlab.com/YottaDB/DB/YDB/-/merge_requests/1132#note_896366838)
which was due to YottaDB using `ESTABLISH` in that file and thus needing a `REVERT`
in each return path. This commit adds a `v63012/missredrevert` test that tests for
this issue.