Temporarily disable r132/ydb704 test on CLANG 14 builds due to valgrind issues
The `r132/ydb704` test fails with error messages like `### unhandled dwarf2 abbrev form code 0x25`
when run on a build that was built with CLANG 14. This is due to a valgrind issue with the new
version of clang. (see https://bugzilla.mozilla.org/show_bug.cgi?id=1758782 for an example of the
issue affecting Firefox). The fix is to temporarily disable the test until the valgrind issue is fixed.

This is done by moving an existing check for the clang version in `com/is_libyottadb_asan_enabled.csh`
outside of the block it is in so that the clang version is set in the test system's environment for
clang builds. This does not set the clang version for gcc builds so the test will continue to run
on gcc builds even if the system's clang version is `14`.