Fix v63012/optimizexecute subtest to handle dynamic literals and disable subtest for PRO builds
The `v63012/optimizexecute` test failed in in-house testing whenever the test was
run without dynamic literals which is a setting randomly chosen by the test system.
This happened because the reference file included the correct output with dynamic
literals but incorrect output without dynamic literals. It is reasonable to expect
that the output would be different with and without dynamic literals so this commit
updates the reference file to handle both. With this change, the test passes reliably.

This commit also disables the `v63012/optimizexecute` test on pro builds because
`yottadb -machine -lis=` only works on dbg builds.