[YottaDB/Lang/YDBGo#42] Change ydbgo34d routine in go/ydbgo34 to better isolate
After YDBGo#42, it was observed that running two or go/ydbgo34 tests at the same time resulted in collisions
that caused the test to fail. The test is waiting for a SIGACKTIMEOUT message to appear in the syslog but when
multiple instances of the test are running, sometimes more than one occurs causing the test to fail. This change
adds some pid checking so we only select the message from THIS run and don't collide with other instances of
the test.