[#419] [YDB#864] New v63012/gtm9182 and r136/ydb864 tests that a MUPIP BACKUP to a backfile file path > 255 characters returns a FILENAMETOOLONG error
This commit adds a `v63012/gtm9182` test that checks that performing a `MUPIP BACKUP` to
a backfile file path > 255 characters will return a `FILENAMETOOLONG` error instead of
successfully backing the file up to the excessively long path. It also adds a `r136/ydb864`
test that runs both an online backup and a `-noonline` backup on path lengths between 220
and 265 checking for the appropriate error messages.