[YDB#861] Test $ZATRANSFORM() invocations that were previously giving wrong answers or SIGSEGV
This test does some nested loops to generate a series of $ZATRANSFORM() invocations. The values
used include those that were noted to be wrong in YDB#861 but include other types of values to
expand the testing a little bit. We have a mix of literals, and expressions.

The two fields we are primarily interested in testing with this test are the following:
- The first field - primary issue is with expressions that the compiler does not optimize (like 1&0)
  but we varied this a bit.
- The third field needed to be either -2 or 2 to invoke the sequential collation char to the first
  char of the input string.
- We vary the 4th field between 0 and 1 to invoke both numeric and string processing.

Note r1.30 (and GT.M through V70002) give a different result for the first four tests done. For
these tests, the computed input value is 0 so should be the same as specifying a 0 there. But doing
that gives different results proving these values with unoptimized expressions are incorrect in
prior versions. In r1.30, it is because $ZATRANSFORM() does not support 2/-2 in the 3rd argument.
It treats these (invalid) values as if '1' was specified.