Fix previous commit that didn't disable valgrind tests on clang 14
This commit fixes an issue with commit 3b67cd6e99424b072cfe81012cfa3d0e7f911f62 which
was supposed to disable valgrind tests when YottaDB was built with clang 14 but it
didn't. The issue was in the command to set the clang version. A print statement in
the r132 inref file got the value of `$clangmajorversion`:

```
14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14
```

The fix is to add an additional `gawk` to the `set clangmajorversion` command which
ensures that `$clangmajorversion` is set to only the first `14` (or whatever the
version is). With this fix, the `r132/ydb704` test (the valgrind test) no longer
runs on the clang14 build.

This commit also fixes an error in `com/outref.awk` that caused the `CLANG14`
option for `##SUSPEND_OUTPUT/##ALLOW_OUTPUT` to not work correctly. The
option was checking the value of `$clang_major_version` but the variable is
called `$clangmajorversion`. This is now fixed and the test passes.