Fix r130/ydb534 due likely to changes in glibc changing the nesting level in a core trace that is in the reference file
Previously we had to allow levels 7 or 8 but now several more levels have been added to a abort() call in the test program so
that we need to allow level 11 and 12 to show up on the do_abort() line. Interestingly, the stack is not deterministic so we
had to add both 11 and 12 to the list of allowed values. We note a lot of new glibc entry points being called as part of
delivering the signal to us.

Note earlier failures on our Ubuntu 22.04 Ryzen 7 box were also getting a core from gdb when it looked at the core that was
intentionally caused by the abort() the test does. We do not know why this started happening nor do we know why it suddenly
stopped happening a few mins ago. I've started a run of 50 tests to verify it isn't just a fluke.