Fix r136/ydb861 subtest failure when UTF-8 chset is used
The forms of $ZATRANSFORM() used in the test only use the values -2 and 2 for the 3rd parameter.
This 3rd parameter defines what type of transformation is requested. The -2 and 2 values mean for
$ZATRANSFORM() to take the first character of the input value (1st parm) and return either the
previous character or the following character respectively.

This form of $ZATRANSFORM() is documented as not supported with UTF-8 so it only deals with ASCII
characters. The reference file has been modified to still pass the test if UTF-8 mode is randomly
selected since making sure it doesn't sig-11 was one of the reasons for this change.