[#416/GTM-7759] Add sudo/gtm7759 test to test the restrict facility with the LOGDENIALS feature
This test is verifying the function of the LOGDENIALS statement in Restrict Facility. This facility controls
when/if certain errors (those so marked in the error source files .. e.g. merrors.msg) are automatically
sent to the syslog. If there is no $ydb_dist/restrict.txt, all thusly marked errors are automatically sent
to the syslog for all users. If the restrict.txt file exists and it has a LOGDENIALS then only the group
mentioned in the LOGDENIALS statement has its messages logged. The userids of other groups do not have these
messages logged.

This test creates a group and userid on the fly since we run the test twice - once under the userid running
the test and again under another id with a different group. Because we are dealing with multiple users, this
test is located in the sudo suite and we need this capability to create the userids we need to create as well
as allowing access to files in this test running under one uid to be made available to another user and group
that we spawn.

This test drives 3 different messages (DBPRIVERR, SETZDIR, and ZGBLDIRACC) and runs each one under each of the
userids. We also create a full copy of $ydb_dist so we can install a restrict.txt file in it for our test. The
file contains only a LOGDENIALS statement for ydbtest3.