Fix outref issue in v63012 test that causes failure on pro builds
A typo in 67b8ef58f0a438d404d30c12003a4660cdb617e0 caused the v63012 test to fail on pro builds on some machines. The
typo was in `v63012/outref/outref.txt` and involved an `##ALLOW_OUTPUT_PRO` (with `_` between `OUTPUT` and `PRO`) when
it should have been `ALLOW_OUTPUT PRO` (with a space). This caused the `##ALLOW_OUTPUT` to be ignored by the test
system which meant that `v63012 test DONE.` was not expected on pro builds. When this line ended up in the output, the
test would fail even though all of the subtests had passed. This commit fixes the typo.