Clearer failure output for sudo/plugins test
When a plugin install fails in the `sudo/plugins` test, the test first prints out
the line `ydbinstall returned a non-zero status: $status` then cats the contents of
the file `install.out`. For the parts of the test that test that `--plugins-only`
option, it also cats the contents of `plugins.out` following `install.out`. Since
the first line of `install.out` will be something along the lines of
`YottaDB version r### installed successfully`, this is confusing. This commit
adds lines clarifying that the contents of `install.out` and `plugins.out` follow.
In addition, the existing code was redundant since the test would cat `install.out`
whether it was `--plugins-only` or not so the if block that cats `install.out` and
`plugins.out` has been simplified.