[#434] Set $SHELL to /usr/local/bin/tcsh at the beginning of submit.awk
Setting $SHELL to `/usr/local/bin/tcsh` in `com/submit.awk` since tcsh is the only supported shell for the test system and this way the user doesn't have to change their default/login shell.

The initial fix set $SHELL in `com/submit_test.csh`, which is invoked once for each test in the set of submitted tests (e.g., in a E_ALL there can be 235 tests) which is unnecessary. Therefore the final fix is to set this env var in `com/submit.awk` where we set this env var once as part of the script that starts all tests one by one.