Disable v63012/gtm9260 subtest from being run on ARMV6L
This test is trying to prove that when YottaDB creates a new shared memory area to hold
an expanded M lock hashtable and that process terminates without cleanup, that a MUPIP RUNDOWN
cleans up the extended hashtable segment (which it did not do prior to V63012).

Getting YottaDB to create this expanded M lock hashtable segment is not straightforward.
There has to be a lot of locks and various parameters have to be "just right". It took quite
awhile to find a blend of lock space size, lock name size, and the number of hash bits used
(we use $ydb_lockhash_n_bits to limit hash uniqueness). Through a long trial and error period,
we discovered how to make the segment appear with 64 bit architectures.

The problem on ARMV6L, is these boxes are 32 bit which changes the size of most of the lock
related control blocks plus it reduces the size of the hash from 64 to 32 bits so the rate
of block consumption is totally different. Rather than spend a bunch of time making this test
function on 32 bits, this fix disables gtm9260 on ARMV6L.