[#415/GTM-5381] Add v63012/gtm5381 test to validate FULLBLKWRT functionality
Test creates straces of a database updates into an empty DB and checks the pwrite64() DB calls
for the max write size which is indicative of FULLBLKWRT functionality. This is done for each
of the FULLBLKWRT settings (0,1,2).

Note - in the reference file, needed to wildcard one of the block writes when FULLBLKWRT=0.
Depending on what random options were chosen, this write size alternated between 28 and 32 bytes.