Fix random r136/ydb864 test failures due to mupip backup processing multiple regions in a non-deterministic order
The r136/ydb864 test failed on 4 machines due to an issue with the reference
file involving the `-noonline` backups for lengths 234 to 236. The issue is
that the test has 2 different regions `DEFAULT` and `AREG` and attempts to
do a backup on both regions. If the test decides to back up `AREG` first,
this backup is successful because the path to the `AREG` temp file is under
the 255 character limit. However, when it backs up `DEFAULT` for those lengths,
the backup fails with an error message. The reference file was set up to expect
`AREG` to be backed up first followed by `DEFAULT`. However, if MUPIP attempted
to back up `DEFAULT` first, 2 lines of output in the reference file would be
missing. Both outputs are correct in terms of the code but since it isn't
possible to do a 3 line `##TEST_AWK`, the fix was to do a trial backup and swap
the databases if the `mumps.dat` database for the `DEFAULT` region is backed
up first. Since the order is set in the inode when the databases are created,
swapping the databases ensures that the database for `AREG` is always backed
up first which ensures a deterministic output.

Since the order of the backups is now known, the reference file has been updated
to remove `##TEST_AWK` usages and regexes that allowed the test to accept either
order. With these changes, the test now passes reliably.