[YDB#487/YDBTest#418/GTM-9157] Add v63012/gtm9157 to test new robustness when repl source server DNS lookup fails and added detail to error
This issue added support to the replication source server so that if the targeted secondary system name fails to resolve,
instead of shutting down as happened prior to V63012, the server keeps running retrying the connection until it resolves.
The purpose being to be able to withstand temporary downs due to network reconfigurations or other temporary issues.
Additionally, some extra infor associated with the GETADDRINFO message put in the source server log contains the name and
port we are trying to resolve.

Note this MR has a dependency on YDB#879 and is also a test for it. YDB#879 contains fixes for things discovered while this
test for GTM-9157 was being developed. The issues fixed by YDB#879 are:
  1. Inability to reliably shutdown the source server at the end of the test. The shutdown usually times out and leaves
     the server spinning.
  2. When this test passes a bogus secondary name to attach to, the source server goes into a hard spin and ignored all
     CONNECTPARAMS settings.
  3. The "time waiting to connect" type messages in the source server log were displaying VERY wrong values. Example, I
     saw one log where the source server had been waiting over 40K seconds but had been running less than 2 minutes.
These issues were all related. Previous to V63012, a failure to resolve a name was fatal to the source server. Now it is
not but in the paths where an error occurred, we still need to sleep the requisite amount for the attempt and we need
to check if any interrupts have come in in order to be able to check if, for example, a source server shutdown has been
requested.

So this test tests all of these things above with one exception - we have no good way to verify that the wait durations
on the source server log messages is correct. So that is not tested directly though after several runs with the -KEEP
test option, we verified that the logs are correct now for both hard and soft spins. The fact is normally, the test
runs fast enough (on the x86 boxes) to request a shutdown partway through the hardspin loop. But there is a randomization
in the test so that 30% of the time, it will induce a 10 second sleep before driving the source server stop which moves
the stop into the soft retry loop.