[YDB#604] [V63013] Update reference files for v63000/gtm7762 and v63006/gtm8947 to reflect V6.3-013 changes
The `v63006/gtm8947` test started failing due to a change in V6.3-013 that now computes the result
of a $translate command at compile time. Thus, we no longer expect an `OC_FNTRANSLATE_FAST` from
the $translate. This commit updates the reference file accordingly.

The `v63000/gtm7762` test started failing due to a change that now returns a `YDB-W-BADCHAR` and
a `YDB-E-BADCHAR` for the line `set f=$TRANSLATE("Hello, this is doge?"_$ZCH(194),"Htid","DdiD")`
due to a change to `sr_port/f_translate.c` to now call `valid_utf_string()`. This `valid_utf_string()`
call returns a `YDB-W-BADCHAR` and then the command later fails with a `YDB-E-BADCHAR`. Previously,
it only produced the `YDB-E-BADCHAR`. The fix is to update the reference file to reflect this change.