[YDB#604] [V63013] Force v53004/C9905001119A test to use BG due to V6.3-013 change that breaks NOISOLATION with MM
After integrated the V6.3-013 changes into YottaDB, the `v53004/C9905001119A` test failed whenever it
was run with the MM access method. This was due to the V6.3-013 changes that removed the `NOISOLATION`
optimization for MM (see https://gitlab.com/YottaDB/DB/YDB/-/merge_requests/1166#note_951687754). As
this test verified that there were less than 5% TP restarts when using NOISOLATION, which was dependent
upon the removed optimizations, the test would always fail with MM. This failure is fixed by forcing
the test to use BG as the access method. With this change, the test now passes reliably.