[YDB#604] [V63013] Update unicode/functions test to reflect GTM-9295 changes to $translate
After integrating V6.3-013 into YottaDB, the `unicode/functions` test started failing
due to multi-byte characters being translated in UTF-8 mode. Previously, they were
not translated for a `$translate` but were translated by a `$ztranslate`. This upstream
change was intentional as the release note GTM-9295 states:

```
In addition, UTF-8 mode $TRANSLATE() evaluations with NOBADCHAR checking give uniform
results, typically matching the $ZTRANSLATE() result with the same arguments. Previously
the results depended on the code path followed to create the string containing the bad
character(s).
```

Thus, the correct fix for this was to update the M code for the test to expect these
multi-byte bad characters to be translated. With this update, the test now passes
reliably.