[expect/gtm8223] Add expect debug logic to this test which failed oddly
We've done thousands of reruns (11 hrs of all systems running debug version and 16 hrs
of all systems running pro version) all without any failures so whatever happened is
VERY difficult to reproduce. But note that those runs were run with this MR in place and
this MR adds some light debugging to expect so it records its actions. This bit of extra
overhead may be all that was needed to change the timing of this test to prevent this failure
from occurring.

Here is the original failure diff:
```
16c16
< PASS from gtm8223
---
> FAIL from gtm8223. Please check gtm8223/gtm8223.diff and gtm8223/errs_found.logx for errors in time sorted fashion
--------- Subtest level diff follows ---------

gtm8223 took 00:00:06:01 from 2022.06.04.06.26.28.EDT to 2022.06.04.06.32.29.EDT : 920 FAIL
--- gtm8223/gtm8223.diff ---
9a10,12
> /extra1/testarea1/brad/V999/tst_V999_R135_dbg_30_220603_220819/expect_0/gtm8223/gtm8223.out
> %YDB-E-GVUNDEF, Global variable undefined: ^%GI
> %YDB-E-INVCMD, Invalid command keyword encountered
```

Here is the original analysis of the issue:
```
I do see what this issue is but am not sure exactly what is causing it or how to fix it. If it were to occur and
generate a debug file and a "sanitized" version of the output file (which I am creating an MR to do), it might be
easier to see where/why it went off the rails. Here's the code driving expect at the time:

    send -- "do ^%GI\r"
    expect "Input device: <terminal>:"
    send -- "\003"                      <== Here a ^C is sent
    expect "YDB>"
    send -- "do ^%GI\r"

And here's the result of that code from the (newly) sanitized expect output file:

    YDB>do ^%GI

    Global Input Utility

    Input device: <terminal>: d        <== Here a ^C is sent
    YDB>o ^%GI
    %YDB-E-GVUNDEF, Global variable undefined: ^%GI
    YDB>gtm8223a.go
    %YDB-E-INVCMD, Invalid command keyword encountered
            gtm8223a.go
            ^-----

Note where the ^C is sent, part of the next command input (the 'd' from the DO) got applied as "input device" so the
next command was totally messed up and executed as an OPEN command instead of a DO command. This looks to me like an
expect issue but a temporary (or permanent?) workaround may be adding a short sleep after sending a ^C to be sure ^C
processing is complete before sending the next command (as distasteful as adding sleeps sounds). I would have thought
that would be enforced by expect 'expecting' the 'YDB>' prompt but it does not appear that happened.

```