[#444] New v63013/gtm9147 test for larger maximum journal buffer size
This commit adds a new `v63013/gtm9147` test that provides coverage
for journal buffer sizes between 32769 and 1048577 blocks to reflect
the new limit of 1048576 blocks. This is very similar to the existing
test `set_jnl/qualifiers_with_args` which covers values from 2307 to
32769 reflecting the old limit of 32768. The difference being that
this test covers the now allowed values that are allowed in V6.3-013+
that were not allowed in V6.3-012 and earlier.