Update r136/ydb864 test to use white box test where needed
This commit updates the `r136/ydb864` test to use the new white box test
`WBTEST_YDB_STATICPID` for the `-noonline` portion of the test. This
addresses a test failure where the test could unexpectedly succeed for
the "DEFAULT" region with a path of 234 characters if the pid was 9999
or less. The white box test ensures that the process ID part of the temp
file name is always 99999 which ensures that the length of the temp file
will be as expected. Since white box test cases only work on dbg builds,
this commit also disables the `-noonline` part of the test on pro builds
and updates the reference file accordingly.