[YDB#604] [V63013] Fix v62000/gtm8145 failure on RHEL by changing terminal type from vt320 to ansi
The v62000/gtm8145 test started failing on the in-house RHEL machines. The test runs an
expect script that tests how ZHELP interacts with CTRL-C commands. After integrating
V6.3-013 into YottaDB, this test started failing on RHEL machines with `YDB-I-NOTERMENTRY`
messages. Upon investigation, this was due to the lack of `vt320` on the RHEL machines.
The fix for this is to change the terminal type to `ansi` which is present on all machines.